изменено:      apk/install_v2ray.sh

+++ apk/install_v2ray.sh
@@ -12,9 +12,11 @@ then
      su - root -c  "pm install /storage/emulated/0/Download/$FILE"
 fi
 
-curl -L -o   com.v2ray.ang.tar.xz     https://github.com/definitly486/Lenovo_Tab_3_7_TB3-730X/releases/download/shared/com.v2ray.ang.tar.xz
+if ! [  -f "com.v2ray.ang.tar.xz" ]; then
+   curl -L -o   com.v2ray.ang.tar.xz     https://github.com/definitly486/Lenovo_Tab_3_7_TB3-730X/releases/download/shared/com.v2ray.ang.tar.xz
+fi
+
 tar xf com.v2ray.ang.tar.xz
-su - root -c "rm -r /data/data/com.v2ray.ang/"
 su - root -c "cp -r  $DIR/com.v2ray.ang /data/data/"
 su - root -c "chmod -R  755 /data/data/com.v2ray.ang/"
 su - root -c "chown  -R  u0_a205:u0_a205 /data/data/com.v2ray.ang/"
